add season selector to stats page

+++ DragonFruit.Six.Client/Screens/Stats/Stats.razor.cs
@@ -23,7 +23,7 @@ namespace DragonFruit.Six.Client.Screens.Stats
 {
     public partial class Stats
     {
-        private const int ModernSeasonStart = 23;
+        private const byte ModernSeasonStart = 23;
         internal const int ModernStatsRange = 60;
 
         private SearchProvider _searchProvider;
